adding NOT_IN_GROUP exercice : code and comments
